pyproject.toml
[tool.poetry]
name = "expense_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^13.14"
python-dotenv = "^0.21.0"
Django = "^4.1.2"
mypy = "0.961"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
Pillow = "^9.2.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
ipython = "^8.5.0"
flake8 = "^5.0.4"
isort = "^5.10.1"
black = "^22.10.0"
django-stubs = {extras = ["compatible-mypy"], version = "^1.12.0"}

[tool.isort]
known_local_folder =["app", "hello", "expense_chat"]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "expense_recorder.settings"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
